+++ mobile/extension.d.ts
@@ -1,11 +1,12 @@
-declare module '*.svg' {
-  import React from 'react';
-  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
-  const src: string;
-  export default src;
-};
+declare module "*.svg" {
+  import React from "react";
+  import { SvgProps } from "react-native-svg";
 
-declare module '*.png' {
+  const content: React.FC<SvgProps>;
+  export default content;
+}
+
+declare module "*.png" {
   const value: any;
   export default value;
-};
+}